requirement.txt
aiohttp
pillow
numpy
fastapi
uvicorn[standard]
orjson
gymnasium
stable_baselines3[extra]
sb3-contrib
simpy
prometheus-api-client
kubernetes
torchvision
datasets
torchaudio
soundfile
librosa
scikit-optimize
optuna
SciencePlots
